feat: Store chat messages in store and fetch them when requested
This commit introduces the functionality to store chat messages in the store.
It also updates the `get_message` to look in the store if message doesn't exist in memory.
This will allow us to retrieve messages in chain.
The head of the conversation is now being stored in store also.
Also, changes the wasm component installation name.